+++ .eslintrc.js
@@ -48,6 +48,7 @@ module.exports = {
       "@typescript-eslint/no-unsafe-call": "off",
       "@typescript-eslint/no-unsafe-member-access": "off",
       "@typescript-eslint/restrict-template-expressions": "off",
+      "@typescript-eslint/no-floating-promises": "off",
       "functional/functional-parameters": "off",
       "functional/prefer-type-literal": "off",
       "functional/no-conditional-statement": ["off"],